backend/resilientlane/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4                # The Django
djangorestframework==3.13.1  # API framework
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
